Separated blacklist view and utils into the subapp, updated docs
